+++ src/main/scala/ajr/rundeck/telegram/TelegramMessenger.scala
@@ -15,11 +15,11 @@ class TelegramMessenger(botAuthKey: String, val baseUrl: String = "https://api.t
     (request.code, request.body)
   }
 
-  def sendMessage(chat: Int, message: String) = {
+  def sendMessage(chat: Long, message: String) = {
     command("sendMessage", Map("chat_id" -> chat.toString, "text" -> message))
   }
   
-  def sendDocument(chat: Int, fileBytes: Array[Byte], fileName: String, mimeType: String = "text/plain") = {
+  def sendDocument(chat: Long, fileBytes: Array[Byte], fileName: String, mimeType: String = "text/plain") = {
     val url = s"$baseUrl/bot$botAuthKey/sendDocument"
     val params = Map("chat_id" -> chat.toString)
     val request = http(url).params(params).postMulti(MultiPart("document", fileName, mimeType, fileBytes)).asString
